adding of stocks

+++ database/migrations/2023_03_30_124000_create_supplies_tables.php
@@ -18,8 +18,8 @@ return new class extends Migration
             $table->string('description');
             $table->integer('stocks')->default(0);
             $table->integer('quantity')->default(0);
-            $table->integer('unit_price')->default(0);
-            $table->integer('total_price')->default(0);
+            $table->float('unit_price')->default(0);
+            $table->float('total_price')->default(0);
             $table->date('date_purchased')->nullable();
             $table->string('item_code')->unique();
             $table->timestamps();
